updated images and styling
