+++ app/(home)/components/SurveySection.tsx
@@ -1,0 +1,9 @@
+import SurveyForm from "../../SurveyForm"
+
+export default function SurveySection() {
+	return (
+		<section className="section bg-white">
+			<SurveyForm />
+		</section>
+	)
+}
