+++ app/components/Header.tsx
@@ -1,7 +1,6 @@
 "use client"
 
 import Link from "next/link"
-import ThemeToggle from "./ThemeToggle"
 import { useEffect, useState } from "react"
 import { FiMoon, FiSun } from "react-icons/fi"
